% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votepath.r
\name{draw_coef}
\alias{draw_coef}
\title{Generate draws from the distribution of coefficients}
\usage{
draw_coef(x, ..., n = 1000)
}
\arguments{
\item{x}{model object whose coefficients are to be drawn}

\item{...}{Other arguments to be passed down}

\item{n}{number of draws to take}
}
\description{
Generate draws from the distribution of coefficients
}
