% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votepath.r
\name{find_type}
\alias{find_type}
\title{Find type of variable}
\usage{
find_type(x, ...)
}
\arguments{
\item{x}{a vector of values to be evaluated}

\item{...}{other arguments to be passed down, currently not implemented.}
}
\description{
Find the variable type to identify the appropriate model for the variable
}
