% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votepath.r
\name{print.summary.simeff}
\alias{print.summary.simeff}
\title{Print method for summary.simeff objects.}
\usage{
\method{print}{summary.simeff}(
  obj,
  ...,
  total_effect = c("calculated", "inferred"),
  direct_effect = c("calculated", "inferred"),
  indirect_effect = c("inferred", "calculated"),
  print_mediated = FALSE
)
}
\arguments{
\item{obj}{Object of class \code{summary.simeff}.}

\item{...}{Other arguments to be passed down, currently unimplemeneted.}
}
\description{
Print method for the output from summary.simeff.
}
