% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votepath.r
\name{prob}
\alias{prob}
\title{Set-up for Drawing Values}
\usage{
prob(obj, b, data, ...)
}
\arguments{
\item{obj}{A model object - currently supported models are \code{lm}, \code{glm}, \code{polr}, \code{multinom}, \code{svyglm}, \code{svyolr} and \code{svymultinom}.}

\item{b}{A list of model coefficients generated by \code{draw_coefs()}.}

\item{data}{A data frame that will be used to generate predictions.}

\item{...}{Other arguments to be passed down, currently none implemented.}
}
\description{
The \code{prob} function sets up values that will be past \code{draw_vals}.  For linear models and
gaussian GLMs, these are simply predicted values.  For non-linear models, these are inverse-link
transformed linear predictor values (e.g., predicted probabilities).
}
