% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votepath.r
\name{summary.simeff}
\alias{summary.simeff}
\title{Summary method for VotePath Simulated Effects}
\usage{
\method{summary}{simeff}(object, ..., conf.level = 0.95)
}
\arguments{
\item{object}{Object of class \code{simeff}}

\item{...}{Other arguments, currently unimplemented}

\item{conf.level}{Level at which to make the confidence interval}
}
\description{
summary method for objects of class \code{simeff}
}
