% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votepath.r
\name{vote_path}
\alias{vote_path}
\title{Specify a Vote Path Analysis}
\usage{
vote_path(blocks, models = NULL, data, ...)
}
\arguments{
\item{blocks}{A list in which each element contains a vector of variable names for variables that will represent
that block's effects.  The list may be named or not.}

\item{models}{If \code{NULL}, the algorithm will use \code{find_type()} to find the appropriate model.  This assumes that variables
that should be estimated using ordinal regression models are ordered factors.  Otherwise, models can be a named vector of
model types for some or all of the variables included in the model.}

\item{data}{A data frame containing all of the variables in the \code{blocks} list.}

\item{...}{Other arguments to be passed down.}
}
\description{
The vote path model estimates independent models that would otherwise use a block-recursive
structure.  Each model in each block is fit independently using the appropriate error distribution.
}
