% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votepath.r
\name{draw_coefs}
\alias{draw_coefs}
\title{Draw coefficients from posterior distribution}
\usage{
draw_coefs(obj, R = 100, ...)
}
\arguments{
\item{obj}{A model object - currently supported models are \code{lm}, \code{glm}, \code{polr}, \code{multinom}, \code{svyglm}, \code{svyolr} and \code{svymultinom}.}
}
\description{
Draw coefficients from their posterior distributions implied by the model.
}
