% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votepath.r
\name{draw_val}
\alias{draw_val}
\title{Take a random draw from a variable variable}
\usage{
draw_val(obj, data, ...)
}
\arguments{
\item{obj}{Model object of class \code{lm}, \code{glm}, \code{polr} or \code{multinom}}

\item{data}{Data frame giving variables used to estimate model}

\item{...}{Other arguments to be passed down, currently not implemented.}

\item{n}{Array of coefficients to be used in the draw.}
}
\description{
Take a random draw from a variable variable
}
