% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votepath.r
\name{draw_val}
\alias{draw_val}
\title{Take a random draw from a variable variable}
\usage{
draw_val(obj, probs, ...)
}
\arguments{
\item{obj}{Model object of class \code{lm}, \code{glm}, \code{polr} or \code{multinom}}

\item{probs}{A list that was produced by the \code{prob()} function.}

\item{...}{Other arguments to be passed down, currently not implemented.}

\item{e}{matrix or vector of errors to add to the predicted values to generate the draw in an \code{lm} or gaussian \code{glm}.}
}
\description{
Take a random draw from a variable variable
}
